[~] Move Organisation - add some optional properties in grid and hint
